+++ src/csvDocumentView.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { window, workspace, WebviewPanel, ExtensionContext, ViewColumn } from 'vscode';
+import { window, workspace, WebviewPanel, ExtensionContext, ViewColumn, WorkspaceEdit, Range, TextDocument, Position } from 'vscode';
 import { URI } from 'vscode-uri';
 import BaseDocumentView from './baseDocumentView';
 import { getLicenseKey } from './license';
@@ -23,6 +23,7 @@ export default class CsvDocumentView extends BaseDocumentView {
     }
 
     public languageId: string;
+    private separator: string;
 
 	public getOptions(): any {
 		let config = workspace.getConfiguration('csv-preview');
@@ -34,6 +35,9 @@ export default class CsvDocumentView extends BaseDocumentView {
             sep = ";";
         } else if (lang === 'csv (pipe)') {
             sep = "\\|";
+        }
+        if (this.separator) {
+            sep = this.separator;
         }
         return {
             separator: sep,
@@ -54,13 +58,117 @@ export default class CsvDocumentView extends BaseDocumentView {
         };
 	}
 
+    private _wsEdit: WorkspaceEdit;
+    private _document: TextDocument;
+    private _currentRange: Range;
+    private _currentRow: string[];
+
+    private editRow(row: number) {
+        let options = this.getOptions();
+        let sep = options.separator;
+        let quote = options.quoteMark;
+        let offset = options.hasHeaders ? 1 : 0;
+        if (this.separator) offset++;
+
+        this._currentRange = new Range(row + offset, 0, row + offset + 1, 0);
+        let line = this._document.getText(this._currentRange);
+
+        // http://markmintoff.com/2013/03/regex-split-by-comma-not-surrounded-by-quotes/
+        let regexItems = new RegExp(`${sep}(?=(?:[^${quote}]*${quote}[^${quote}]*${quote})*[^${quote}]*$)`);
+        this._currentRow = line.split(new RegExp(regexItems));
+    }
+
+    private setCellValue(row: number, col: number, value: string) {
+        let options = this.getOptions();
+        let sep = options.separator;
+        let quote = options.quoteMark;
+
+        if (!this._currentRow) {
+            this.editRow(row);
+        }
+
+        let regexSeparator = new RegExp(sep);
+        this._currentRow[col] = regexSeparator.exec(value) ? `${quote}${value}${quote}` : value;
+    }
+
+    private deleteRows(rows: number[]) {
+        let options = this.getOptions();
+        let offset = options.hasHeaders ? 1 : 0;
+        if (this.separator) offset++;
+        
+        rows.forEach((row: number) => {
+            let lineRange = new Range(row + offset, 0, row + offset + 1, 0);
+            this._wsEdit.delete(this.uri, lineRange);
+        }, this);
+    }
+
+    private appendRow(columns: number) {
+        let options = this.getOptions();
+        let sep = options.separator;
+        let empty = sep.repeat(columns - 1);
+        let row = this._document.lineCount - 1;
+
+        this._currentRange = new Range(row, 0, row, 0);
+        this._currentRow = empty.split(new RegExp(options.separator));
+    }
+
+    private beginEdit() {
+        if (!this._wsEdit) {
+            this._wsEdit = new WorkspaceEdit();
+        }
+    }
+
+    private async endEdit(refresh?: boolean) {
+        if (this._wsEdit) {
+            if (this._currentRow) {
+                let sep = this.getOptions().separator;
+                this._wsEdit.replace(this.uri, this._currentRange, this._currentRow.join(sep));
+            }
+            await workspace.applyEdit(this._wsEdit);
+            this._wsEdit = undefined;
+            this._currentRange = undefined;
+            this._currentRow = undefined;
+            if (refresh) {
+                this.refresh();
+            }
+        }
+    }
+
+    public enableEditing(document: TextDocument) {
+        this._document = document;
+        this.webview.onDidReceiveMessage((e) => {
+            if (e.cellEditEnded) {
+                this.beginEdit();
+                this.setCellValue(e.row, e.col, e.value);
+            }
+            else if (e.rowEditEnded) {
+                this.endEdit();
+            }
+            else if (e.deleteRows) {
+                this.beginEdit();
+                this.deleteRows(e.rows);
+                this.endEdit();
+            }
+            else if (e.separator) {
+                this.separator = e.separator;
+            }
+            else if (e.rowAdded) {
+                this.appendRow(e.count);
+            }
+        }, this, this._disposables);
+    }
+
     refresh(): void {
+        if (this._wsEdit)
+            return;
         let self = this;
         workspace.openTextDocument(this.uri).then(document => {
-            self.webview.postMessage({
-                refresh: true,
-                content: document.getText()
-            })
+            if (!self._disposed) {
+                self.webview.postMessage({
+                    refresh: true,
+                    content: document.getText()
+                });
+            }
         }, reason => {
             window.showInformationMessage(reason);
         });