+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 4.0.46 Pre-Release (January 4, 2022)
+Added support for editing CSV files: modifying individual cells, deleting rows, and appending new rows.
+
+Added support for editing Excel files: modifying individual cells, inserting rows and columns, and deleting rows and columns.
+
 ## 4.0.45 (November 30, 2021)
 Added read-only custom editor support. For Excel files, this is the default, and clicking the name of an Excel file in Explorer view opens the custom editor directly. For CSV files, this is optional, and executing the `Open With` command on the context menu prompts for the built-in or custom editor to be opened. The `Open Preview` command is still supported for both file types.